net/bnxt/tf_ulp: support Thor2 ULP layer
This patch includes the support for following features that enable
Thor2 support in the ULP layer:

1. Added support for ulp initialization on Thor2 platform. This involved
breaking the functionality that is common and not common between Thor and
Thor2 platforms.
2. MPC support for Thor2. This feature enables the access of the DRAM
memory location in the HOST CPU for Exact match flows and Action records
for those flows.
3. Added support for VF's on Thor2 platform.
4. Added support to offload traffic between two VF's on the system.
5. Renamed all BNXT_TF_DBG macros to BNXT_DRV_DBG.
6. Added logic to get error conditions in the flow create path.
7. Added support for Geneve header and set TTL action parsing.
8. Add mpc batching to ulp flow create for Thor2.

This patch also updates the template files for the changes
that are being added in this patch.

+++ drivers/net/bnxt/tf_ulp/ulp_flow_db.h
@@ -260,7 +260,7 @@ ulp_flow_db_validate_flow_func(struct bnxt_ulp_context *ulp_ctx,
 int32_t
 ulp_default_flow_db_cfa_action_get(struct bnxt_ulp_context *ulp_ctx,
 				   uint32_t flow_id,
-				   uint16_t *cfa_action);
+				   uint32_t *cfa_action);
 
 /*
  * Set or reset the parent flow in the parent-child database
@@ -413,9 +413,11 @@ void ulp_flow_db_shared_session_set(struct ulp_flow_db_res_params *res,
 				    enum bnxt_ulp_session_type s_type);
 
 /*
- * Get the shared bit for the flow db entry
+ * get the shared bit for the flow db entry
  *
- * res [out] Shared session type
+ * res [in] Ptr to fdb entry
+ *
+ * returns session type
  */
 enum bnxt_ulp_session_type
 ulp_flow_db_shared_session_get(struct ulp_flow_db_res_params *res);
